+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added
   - New linters
     - `oksh` for shell files
+    - `hush --check` for `*.hush` files
     - `actionlint` for GitHub Action workflow files
 
 ## \[0.6.4\] - 2023-07-14